+++ docs/src/demos/components/index.ts
@@ -52,3 +52,5 @@ export * from './Tooltip';
 export * from './VisuallyHidden';
 export * from './XLoader';
 export * from './Calendar';
+export * from './Tag';
+export * from './DateField';